# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gigabee decentralized AI inference network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workers/register:
    post:
      operationId: registerWorker
      tags: [workers]
      summary: Register a new worker node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerRegistration"
      responses:
        "201":
          description: Worker registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workers/me:
    get:
      operationId: getMyWorker
      tags: [workers]
      summary: Get current user's worker status
      responses:
        "200":
          description: Worker status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          description: No worker registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workers/heartbeat:
    post:
      operationId: workerHeartbeat
      tags: [workers]
      summary: Send worker heartbeat to stay online
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatData"
      responses:
        "200":
          description: Heartbeat acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    WorkerRegistration:
      type: object
      required: [type, capabilities]
      properties:
        type:
          type: string
          enum: [browser, native]
        capabilities:
          $ref: "#/components/schemas/WorkerCapabilities"
    WorkerCapabilities:
      type: object
      required: [model, tokensPerSecond]
      properties:
        model:
          type: string
        tokensPerSecond:
          type: number
        vramMb:
          type: ["integer", "null"]
        gpuName:
          type: ["string", "null"]
    Worker:
      type: object
      required: [id, type, status, reputationScore, totalJobsServed, totalTokensGenerated, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [browser, native]
        status:
          type: string
          enum: [idle, serving, offline]
        reputationScore:
          type: number
        totalJobsServed:
          type: integer
        totalTokensGenerated:
          type: integer
        capabilities:
          $ref: "#/components/schemas/WorkerCapabilities"
        createdAt:
          type: string
        lastSeen:
          type: ["string", "null"]
    HeartbeatData:
      type: object
      required: [workerId, status]
      properties:
        workerId:
          type: string
        status:
          type: string
          enum: [idle, serving]
...